create variables file
